Add argument parser
Per the TODO at the top of the file, add a real argument parser to
oe-stylize.  This allows for correct help strings (with --help) as well
as processing of multiple files, either via a folder, in which case said
meta-layer will be searched for bb and bbappend files to process, or
individual filepaths, which for single files, function as they did
before, or for multiple files, allow parsing multiple files at a single
invocation.

Signed-off-by: Ed Tanous <edtanous@google.com>