Add write in place support to oe-stylize
Similar to other tools like clang-format, it is useful to be able to
directly write back the results of an operation to the files in
question.  This implements that feature via an --inplace argument added
to the parser, which when present, will rewrite the result of any files
parsed back to the files themselves.

Signed-off-by: Ed Tanous <edtanous@google.com>