Verify an added newline isn't the first line
There are some cases where bbappend files might not have a SUMMARY,
SRC_URI or other section, and might only include one key being set.  In
this case, we don't want to add a newline at the beginning, because
there was no previous section, and adding a newline would be
nonsensical.

This commit adds a minor check for that condition, and avoids adding the
newline if this is the first segment.

Signed-off-by: Ed Tanous <edtanous@google.com>